refactor: Composable 함수 파일로 분리 및 부모 레이아웃에서 Modifier 적용하게끔 변경
