+++ app/src/main/java/woowacourse/payments/MainActivity.kt
@@ -20,7 +20,7 @@ class MainActivity : ComponentActivity() {
         enableEdgeToEdge()
         setContent {
             AndroidpaymentsTheme {
-                AddCardScreen(Modifier)
+                AddCardScreen()
             }
         }
     }
