+++ app/src/main/java/woowacourse/payments/ui/addcard/AddCardScreen.kt
@@ -27,10 +27,14 @@ import woowacourse.payments.ui.addcard.util.CardNumberTransformation
 import woowacourse.payments.ui.addcard.util.ExpirationDateTransformation
 import woowacourse.payments.ui.addcard.util.PlaceholderTransformation
 import woowacourse.payments.ui.component.Card
+import woowacourse.payments.ui.component.CardNumberTextField
+import woowacourse.payments.ui.component.ExpireDateTextField
+import woowacourse.payments.ui.component.OwnerNameTextField
+import woowacourse.payments.ui.component.PasswordTextField
 import woowacourse.payments.ui.theme.AndroidpaymentsTheme
 
 @Composable
-fun AddCardScreen(modifier: Modifier) {
+fun AddCardScreen() {
     Scaffold(
         topBar = {
             AddCardTopbar()
@@ -38,7 +42,7 @@ fun AddCardScreen(modifier: Modifier) {
     ) { padding ->
         var cardInfo by remember { mutableStateOf(CardInfoUiState()) }
         Column(
-            modifier = modifier
+            modifier = Modifier
                 .padding(padding)
                 .padding(top = 14.dp)
                 .padding(horizontal = 24.dp)
@@ -46,130 +50,38 @@ fun AddCardScreen(modifier: Modifier) {
             verticalArrangement = Arrangement.Center,
             horizontalAlignment = Alignment.CenterHorizontally
         ) {
-            Card(modifier)
-            AddCardContent(modifier, cardInfo)
+            Card()
+            Column {
+                CardNumberTextField(
+                    Modifier
+                        .padding(top = 40.dp)
+                        .fillMaxWidth(),
+                    cardInfo
+                )
+                ExpireDateTextField(
+                    Modifier
+                        .padding(top = 18.dp)
+                        .fillMaxWidth(0.47f),
+                    cardInfo
+                )
+                OwnerNameTextField(
+                    Modifier
+                        .fillMaxWidth(),
+                    cardInfo
+                )
+                PasswordTextField(
+                    Modifier.fillMaxWidth(0.47f),
+                    cardInfo
+                )
+            }
         }
     }
 }
-
-@Composable
-fun AddCardContent(
-    modifier: Modifier,
-    cardInfo: CardInfoUiState,
-) {
-    Column {
-        CardNumberTextField(modifier, cardInfo)
-        ExpireDateTextField(modifier, cardInfo)
-        OwnerNameTextField(modifier, cardInfo)
-        PasswordTextField(modifier, cardInfo)
-    }
-}
-
-@Composable
-fun CardNumberTextField(
-    modifier: Modifier,
-    cardInfo: CardInfoUiState,
-) {
-    OutlinedTextField(
-        modifier = modifier
-            .padding(top = 40.dp)
-            .fillMaxWidth(),
-        value = cardInfo.cardNumber,
-        onValueChange = {
-            cardInfo.onValueChanged(cardNumber = it)
-        },
-        singleLine = true,
-        label = { Text(stringResource(R.string.addcard_card_number_label)) },
-        visualTransformation = if (cardInfo.cardNumber.isEmpty()) PlaceholderTransformation(
-            placeholder = stringResource(R.string.addcard_card_number_placeholder),
-            textColor = colorResource(R.color.payments_placeholder_color)
-        ) else CardNumberTransformation(),
-    )
-}
-
-@Composable
-fun ExpireDateTextField(
-    modifier: Modifier,
-    cardInfo: CardInfoUiState,
-) {
-    OutlinedTextField(
-        modifier = modifier
-            .padding(top = 18.dp)
-            .fillMaxWidth(0.47f),
-        singleLine = true,
-        value = cardInfo.expireDate,
-        onValueChange = {
-            cardInfo.onValueChanged(expireDate = it)
-        },
-        isError = !cardInfo.isExpirationDateValid,
-        label = { Text(stringResource(R.string.addcard_expire_date_label)) },
-        supportingText = {
-            if (!cardInfo.isExpirationDateValid) {
-                Text("유효하지 않은 날짜입니다")
-            }
-        },
-        visualTransformation = if (cardInfo.expireDate.isEmpty()) PlaceholderTransformation(
-            placeholder = stringResource(R.string.addcard_expire_date_placeholder),
-            textColor = colorResource(R.color.payments_placeholder_color)
-        ) else ExpirationDateTransformation(),
-    )
-}
-
-@Composable
-fun OwnerNameTextField(
-    modifier: Modifier,
-    cardInfo: CardInfoUiState,
-) {
-    OutlinedTextField(
-        modifier = modifier
-            .padding(top = 18.dp)
-            .fillMaxWidth(),
-        value = cardInfo.ownerName,
-        onValueChange = {
-            cardInfo.onValueChanged(ownerName = it)
-        },
-        singleLine = true,
-        label = { Text(stringResource(R.string.addcard_owner_name_label)) },
-        supportingText = {
-            Text(
-                modifier = modifier.fillMaxWidth(),
-                textAlign = TextAlign.End,
-                text = "${cardInfo.ownerName.length}/${CardInfoUiState.OWNER_NAME_MAX_SIZE}"
-            )
-        },
-        visualTransformation = if (cardInfo.ownerName.isEmpty()) PlaceholderTransformation(
-            placeholder = stringResource(R.string.addcard_owner_name_placeholder),
-            textColor = colorResource(R.color.payments_placeholder_color)
-        ) else VisualTransformation.None
-    )
-}
-
-@Composable
-fun PasswordTextField(
-    modifier: Modifier,
-    cardInfo: CardInfoUiState,
-) {
-    OutlinedTextField(
-        modifier = modifier
-            .fillMaxWidth(0.47f),
-        singleLine = true,
-        value = cardInfo.password,
-        onValueChange = {
-            cardInfo.onValueChanged(password = it)
-        },
-        label = { Text(stringResource(R.string.addcard_password_label)) },
-        visualTransformation = if (cardInfo.password.isEmpty()) PlaceholderTransformation(
-            placeholder = stringResource(R.string.addcard_password_placeholder),
-            textColor = colorResource(R.color.payments_placeholder_color)
-        ) else PasswordVisualTransformation()
-    )
-}
-
 
 @Preview(showBackground = true)
 @Composable
 private fun AddCardScreenPreview() {
     AndroidpaymentsTheme {
-        AddCardScreen(Modifier)
+        AddCardScreen()
     }
 }
