+++ app/src/main/java/woowacourse/payments/ui/component/CardNumberTextField.kt
@@ -1,0 +1,32 @@
+package woowacourse.payments.ui.component
+
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.colorResource
+import androidx.compose.ui.res.stringResource
+import woowacourse.payments.R
+import woowacourse.payments.ui.addcard.CardInfoUiState
+import woowacourse.payments.ui.addcard.util.CardNumberTransformation
+import woowacourse.payments.ui.addcard.util.PlaceholderTransformation
+
+@Composable
+fun CardNumberTextField(
+    modifier: Modifier = Modifier,
+    cardInfo: CardInfoUiState,
+) {
+    OutlinedTextField(
+        modifier = modifier,
+        value = cardInfo.cardNumber,
+        onValueChange = {
+            cardInfo.onValueChanged(cardNumber = it)
+        },
+        singleLine = true,
+        label = { Text(stringResource(R.string.addcard_card_number_label)) },
+        visualTransformation = if (cardInfo.cardNumber.isEmpty()) PlaceholderTransformation(
+            placeholder = stringResource(R.string.addcard_card_number_placeholder),
+            textColor = colorResource(R.color.payments_placeholder_color)
+        ) else CardNumberTransformation(),
+    )
+}
